add[FE]: Add status to following profile image(#29)
- Add ProfileImage, ProfileImagWithStatus components
- Refactor Following component
