+++ frontend/src/assets/styles/Following.module.css
@@ -4,10 +4,5 @@
 
 .li {
   margin-bottom: 16px;
-}
-
-.img {
-  max-width: 52px;
-  max-height: 52px;
-  border-radius: 50%;
+  display: flex;
 }
