--- frontend/src/assets/styles/Profile.module.css
@@ -1,4 +1,0 @@
-.img {
-  max-width: 320px;
-  max-height: 320px;
-}
