+++ frontend/src/assets/styles/ProfileImage.module.css
@@ -1,0 +1,3 @@
+.img {
+  border-radius: 50%;
+}
