+++ frontend/src/assets/styles/ProfileImageWithStatus.module.css
@@ -1,0 +1,10 @@
+.container {
+  position: relative;
+}
+
+.status {
+  position: absolute;
+  bottom: 0;
+  right: 0;
+  background-color: white;
+}
