+++ frontend/src/components/atoms/ProfileImage.tsx
@@ -1,0 +1,19 @@
+import styles from 'assets/styles/ProfileImage.module.css';
+
+interface Props {
+  profileImageUrl: string | undefined;
+  alt: string;
+  size: number;
+}
+
+export default function ProfileImage({ profileImageUrl, alt, size }: Props) {
+  return (
+    <img
+      src={profileImageUrl}
+      alt={alt}
+      width={size}
+      height={size}
+      className={styles.img}
+    />
+  );
+}
